LPS-53736 Extract DDL tests out of the core
