+++ modules/apps/dynamic-data-lists/dynamic-data-lists-test/test/unit/com/liferay/dynamic/data/lists/DDLTestSuite.java
@@ -1,0 +1,34 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.dynamic.data.lists;
+
+import com.liferay.dynamic.data.lists.asset.DDLRecordDDMFormValuesReaderTest;
+import com.liferay.dynamic.data.lists.model.impl.DDLRecordSetImplTest;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+/**
+ * @author Marcellus Tavares
+ */
+@RunWith(Suite.class)
+@SuiteClasses(
+	{
+		DDLRecordDDMFormValuesReaderTest.class, DDLRecordSetImplTest.class
+	}
+)
+public class DDLTestSuite {
+}
