+++ modules/apps/dynamic-data-lists/dynamic-data-lists-test/test/unit/com/liferay/dynamic/data/lists/model/impl/DDLRecordSetImplTest.java
@@ -1,0 +1,433 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.dynamic.data.lists.model.impl;
+
+import com.liferay.dynamic.data.lists.model.DDLRecordSet;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.util.HtmlUtil;
+import com.liferay.portal.kernel.util.LocaleUtil;
+import com.liferay.portal.kernel.util.Props;
+import com.liferay.portal.kernel.util.PropsKeys;
+import com.liferay.portal.kernel.util.PropsUtil;
+import com.liferay.portal.kernel.xml.Document;
+import com.liferay.portal.kernel.xml.Element;
+import com.liferay.portal.kernel.xml.SAXReaderUtil;
+import com.liferay.portal.util.HtmlImpl;
+import com.liferay.portal.xml.SAXReaderImpl;
+import com.liferay.portlet.dynamicdatamapping.io.DDMFormXSDDeserializerImpl;
+import com.liferay.portlet.dynamicdatamapping.io.DDMFormXSDDeserializerUtil;
+import com.liferay.portlet.dynamicdatamapping.io.DDMFormXSDSerializerImpl;
+import com.liferay.portlet.dynamicdatamapping.io.DDMFormXSDSerializerUtil;
+import com.liferay.portlet.dynamicdatamapping.model.DDMForm;
+import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
+import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
+import com.liferay.portlet.dynamicdatamapping.model.DDMTemplate;
+import com.liferay.portlet.dynamicdatamapping.model.DDMTemplateConstants;
+import com.liferay.portlet.dynamicdatamapping.model.LocalizedValue;
+import com.liferay.portlet.dynamicdatamapping.model.impl.DDMStructureImpl;
+import com.liferay.portlet.dynamicdatamapping.model.impl.DDMTemplateImpl;
+import com.liferay.portlet.dynamicdatamapping.service.DDMStructureLocalServiceUtil;
+import com.liferay.portlet.dynamicdatamapping.service.DDMTemplateLocalServiceUtil;
+
+import java.util.HashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Set;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import org.mockito.Matchers;
+import org.mockito.invocation.InvocationOnMock;
+import org.mockito.stubbing.Answer;
+
+import org.powermock.api.mockito.PowerMockito;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
+
+/**
+ * @author Marcellus Tavares
+ */
+@PrepareForTest(
+	{
+		DDMStructureLocalServiceUtil.class, DDMTemplateLocalServiceUtil.class,
+		LocaleUtil.class
+	}
+)
+@RunWith(PowerMockRunner.class)
+public class DDLRecordSetImplTest extends PowerMockito {
+
+	@Before
+	public void setUp() throws Exception {
+		setUpDDMFormXSDDeserializerUtil();
+		setUpDDMFormXSDSerializerUtil();
+		setUpDDMStructureLocalServiceUtil();
+		setUpDDMTemplateLocalServiceUtil();
+		setUpHtmlUtil();
+		setUpLocaleUtil();
+		setUpPropsUtil();
+		setUpSAXReaderUtil();
+	}
+
+	@Test
+	public void testGetDDMStructure() throws Exception {
+		Document document = createDocument("Text 1", "Text 2", "Text 3");
+
+		DDMStructure ddmStructure = createStructure(
+			"Test Structure", "Text 1", "Text 2", "Text 3");
+
+		DDLRecordSet recordSet = createRecordSet(
+			ddmStructure.getStructureId(), "Test Record Set");
+
+		Element rootElement = document.getRootElement();
+
+		Element dynamicElement = rootElement.element("dynamic-element");
+
+		rootElement.remove(dynamicElement);
+
+		DDMTemplate template = createTemplate(
+			RandomTestUtil.randomLong(), "Test Form Template",
+			DDMTemplateConstants.TEMPLATE_MODE_CREATE, document.asXML());
+
+		Set<String> fieldNames = ddmStructure.getFieldNames();
+
+		DDMStructure recordSetDDMStructure = recordSet.getDDMStructure(
+			template.getTemplateId());
+
+		if (fieldNames.equals(recordSetDDMStructure.getFieldNames())) {
+			Assert.fail();
+		}
+
+		recordSetDDMStructure = recordSet.getDDMStructure();
+
+		if (!fieldNames.equals(recordSetDDMStructure.getFieldNames())) {
+			Assert.fail();
+		}
+	}
+
+	protected void addTextDDMFormFields(DDMForm ddmForm, String... fieldNames) {
+		List<DDMFormField> ddmFormFields = ddmForm.getDDMFormFields();
+
+		for (String fieldName : fieldNames) {
+			ddmFormFields.add(createTextDDMFormField(fieldName));
+		}
+	}
+
+	protected Element addTextElement(
+		Element element, String name, String label, boolean localizable) {
+
+		Element dynamicElement = element.addElement("dynamic-element");
+
+		dynamicElement.addAttribute("dataType", "string");
+		dynamicElement.addAttribute("localizable", String.valueOf(localizable));
+		dynamicElement.addAttribute("name", name);
+		dynamicElement.addAttribute("type", "text");
+
+		Element metadataElement = dynamicElement.addElement("meta-data");
+
+		metadataElement.addAttribute("locale", LocaleUtil.toLanguageId(
+			LocaleUtil.US));
+
+		Element entryElement = metadataElement.addElement("entry");
+
+		entryElement.addAttribute("name", "label");
+		entryElement.setText(label);
+
+		return dynamicElement;
+	}
+
+	protected Set<Locale> createAvailableLocales(Locale... locales) {
+		Set<Locale> availableLocales = new LinkedHashSet<>();
+
+		for (Locale locale : locales) {
+			availableLocales.add(locale);
+		}
+
+		return availableLocales;
+	}
+
+	protected DDMForm createDDMForm(
+		Set<Locale> availableLocales, Locale defaultLocale) {
+
+		DDMForm ddmForm = new DDMForm();
+
+		ddmForm.setAvailableLocales(availableLocales);
+		ddmForm.setDefaultLocale(defaultLocale);
+
+		return ddmForm;
+	}
+
+	protected DDMForm createDDMForm(String... fieldNames) {
+		DDMForm ddmForm = createDDMForm(
+			createAvailableLocales(LocaleUtil.US), LocaleUtil.US);
+
+		addTextDDMFormFields(ddmForm, fieldNames);
+
+		return ddmForm;
+	}
+
+	protected Document createDocument(String... fieldNames) {
+		Document document = createEmptyDocument();
+
+		for (String fieldName : fieldNames) {
+			addTextElement(
+				document.getRootElement(), fieldName, fieldName, false);
+		}
+
+		return document;
+	}
+
+	protected Document createEmptyDocument() {
+		Document document = SAXReaderUtil.createDocument();
+
+		Element rootElement = document.addElement("root");
+
+		rootElement.addAttribute("available-locales", "en_US");
+		rootElement.addAttribute("default-locale", "en_US");
+
+		return document;
+	}
+
+	protected DDLRecordSet createRecordSet(long ddmStructureId, String name) {
+		DDLRecordSet recordSet = new DDLRecordSetImpl();
+
+		recordSet.setDDMStructureId(ddmStructureId);
+		recordSet.setName(name);
+
+		return recordSet;
+	}
+
+	protected DDMStructure createStructure(String name, DDMForm ddmForm) {
+		DDMStructure structure = new DDMStructureImpl();
+
+		structure.setStructureId(RandomTestUtil.randomLong());
+		structure.setName(name);
+		structure.setDefinition(DDMFormXSDSerializerUtil.serialize(ddmForm));
+
+		_structures.put(structure.getStructureId(), structure);
+
+		return structure;
+	}
+
+	protected DDMStructure createStructure(String name, String... fieldNames) {
+		DDMForm ddmForm = createDDMForm(fieldNames);
+
+		return createStructure(name, ddmForm);
+	}
+
+	protected DDMTemplate createTemplate(
+		long templateId, String name, String mode, String script) {
+
+		DDMTemplate template = new DDMTemplateImpl();
+
+		template.setTemplateId(templateId);
+		template.setName(name);
+		template.setMode(mode);
+		template.setScript(script);
+
+		_templates.put(template.getTemplateId(), template);
+
+		return template;
+	}
+
+	protected DDMFormField createTextDDMFormField(String name) {
+		DDMFormField ddmFormField = new DDMFormField(name, "text");
+
+		ddmFormField.setDataType("string");
+		ddmFormField.setLocalizable(true);
+
+		LocalizedValue localizedValue = ddmFormField.getLabel();
+
+		localizedValue.addString(LocaleUtil.US, name);
+
+		return ddmFormField;
+	}
+
+	protected DDMStructure getStructure(long structureId) {
+		try {
+			return DDMStructureLocalServiceUtil.getStructure(structureId);
+		}
+		catch (Exception e) {
+			return null;
+		}
+	}
+
+	protected DDMTemplate getTemplate(long templateId) {
+		try {
+			return DDMTemplateLocalServiceUtil.getTemplate(templateId);
+		}
+		catch (Exception e) {
+			return null;
+		}
+	}
+
+	protected void setUpDDMFormXSDDeserializerUtil() {
+		DDMFormXSDDeserializerUtil ddmFormXSDDeserializerUtil =
+			new DDMFormXSDDeserializerUtil();
+
+		ddmFormXSDDeserializerUtil.setDDMFormXSDDeserializer(
+			new DDMFormXSDDeserializerImpl());
+	}
+
+	protected void setUpDDMFormXSDSerializerUtil() {
+		DDMFormXSDSerializerUtil ddmFormXSDSerializerUtil =
+			new DDMFormXSDSerializerUtil();
+
+		ddmFormXSDSerializerUtil.setDDMFormXSDSerializer(
+			new DDMFormXSDSerializerImpl());
+	}
+
+	protected void setUpDDMStructureLocalServiceUtil() {
+		mockStatic(DDMStructureLocalServiceUtil.class);
+
+		when(
+			getStructure(Matchers.anyLong())
+		).then(
+			new Answer<DDMStructure>() {
+
+				@Override
+				public DDMStructure answer(InvocationOnMock invocationOnMock)
+					throws Throwable {
+
+					Object[] args = invocationOnMock.getArguments();
+
+					Long structureId = (Long)args[0];
+
+					return _structures.get(structureId);
+				}
+
+			}
+		);
+	}
+
+	protected void setUpDDMTemplateLocalServiceUtil() {
+		mockStatic(DDMTemplateLocalServiceUtil.class);
+
+		when(
+			getTemplate(Matchers.anyLong())
+		).then(
+			new Answer<DDMTemplate>() {
+
+				@Override
+				public DDMTemplate answer(InvocationOnMock invocationOnMock)
+					throws Throwable {
+
+					Object[] args = invocationOnMock.getArguments();
+
+					Long templateId = (Long)args[0];
+
+					return _templates.get(templateId);
+				}
+
+			}
+		);
+	}
+
+	protected void setUpHtmlUtil() {
+		HtmlUtil htmlUtil = new HtmlUtil();
+
+		htmlUtil.setHtml(new HtmlImpl());
+	}
+
+	protected void setUpLocaleUtil() {
+		mockStatic(LocaleUtil.class);
+
+		when(
+			LocaleUtil.fromLanguageId("en_US")
+		).thenReturn(
+			LocaleUtil.US
+		);
+
+		when(
+			LocaleUtil.fromLanguageId("pt_BR")
+		).thenReturn(
+			LocaleUtil.BRAZIL
+		);
+
+		when(
+			LocaleUtil.toLanguageId(LocaleUtil.US)
+		).thenReturn(
+			"en_US"
+		);
+
+		when(
+			LocaleUtil.toLanguageId(LocaleUtil.BRAZIL)
+		).thenReturn(
+			"pt_BR"
+		);
+
+		when(
+			LocaleUtil.toLanguageIds((Locale[])Matchers.any())
+		).then(
+			new Answer<String[]>() {
+
+				@Override
+				public String[] answer(InvocationOnMock invocationOnMock)
+					throws Throwable {
+
+					Object[] args = invocationOnMock.getArguments();
+
+					Locale[] locales = (Locale[])args[0];
+
+					String[] languageIds = new String[locales.length];
+
+					for (int i = 0; i < locales.length; i++) {
+						languageIds[i] = LocaleUtil.toLanguageId(locales[i]);
+					}
+
+					return languageIds;
+				}
+			}
+		);
+	}
+
+	protected void setUpPropsUtil() throws Exception {
+		Props props = mock(Props.class);
+
+		when(
+			props.get(PropsKeys.DYNAMIC_DATA_MAPPING_IMAGE_EXTENSIONS)
+		).thenReturn(
+			".gif,.jpeg,.jpg,.png"
+		);
+
+		when(
+			props.get(PropsKeys.DYNAMIC_DATA_MAPPING_IMAGE_SMALL_MAX_SIZE)
+		).thenReturn(
+			"51200"
+		);
+
+		when(
+			props.get(PropsKeys.INDEX_DATE_FORMAT_PATTERN)
+		).thenReturn(
+			"yyyyMMddHHmmss"
+		);
+
+		PropsUtil.setProps(props);
+	}
+
+	protected void setUpSAXReaderUtil() {
+		SAXReaderUtil saxReaderUtil = new SAXReaderUtil();
+
+		saxReaderUtil.setSecureSAXReader(new SAXReaderImpl());
+	}
+
+	protected Map<Long, DDMStructure> _structures = new HashMap<>();
+	protected Map<Long, DDMTemplate> _templates = new HashMap<>();
+
+}
