--- portal-impl/test/integration/com/liferay/portlet/dynamicdatalists/search/DDLRecordSearchTest.java
@@ -1,275 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portlet.dynamicdatalists.search;
-
-import com.liferay.portal.kernel.search.Hits;
-import com.liferay.portal.kernel.search.SearchContext;
-import com.liferay.portal.kernel.search.SearchEngine;
-import com.liferay.portal.kernel.search.SearchEngineUtil;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
-import com.liferay.portal.kernel.test.rule.Sync;
-import com.liferay.portal.kernel.test.rule.SynchronousDestinationTestRule;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.SearchContextTestUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
-import com.liferay.portal.model.Group;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.MainServletTestRule;
-import com.liferay.portlet.dynamicdatalists.model.DDLRecordSet;
-import com.liferay.portlet.dynamicdatalists.service.DDLRecordLocalServiceUtil;
-import com.liferay.portlet.dynamicdatalists.util.test.DDLRecordSetTestHelper;
-import com.liferay.portlet.dynamicdatalists.util.test.DDLRecordTestHelper;
-import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
-import com.liferay.portlet.dynamicdatamapping.storage.DDMFormFieldValue;
-import com.liferay.portlet.dynamicdatamapping.storage.DDMFormValues;
-import com.liferay.portlet.dynamicdatamapping.util.test.DDMFormValuesTestUtil;
-import com.liferay.portlet.dynamicdatamapping.util.test.DDMStructureTestHelper;
-
-import org.junit.Assert;
-import org.junit.Assume;
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-
-/**
- * @author Marcellus Tavares
- * @author André de Oliveira
- */
-@Sync
-public class DDLRecordSearchTest {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE,
-			SynchronousDestinationTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		_group = GroupTestUtil.addGroup();
-
-		DDLRecordSet recordSet = addRecordSet();
-
-		_recordTestHelper = new DDLRecordTestHelper(_group, recordSet);
-		_searchContext = getSearchContext(_group, recordSet);
-	}
-
-	@Test
-	public void testBasicSearchWithJustOneTerm() throws Exception {
-		addRecord("Joe Bloggs", "Simple description");
-		addRecord("Bloggs", "Another description example");
-		addRecord(RandomTestUtil.randomString(), RandomTestUtil.randomString());
-
-		assertSearch("example", 1);
-		assertSearch("description", 2);
-	}
-
-	@Test
-	public void testExactPhrase() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord("Joe Bloggs", "Simple description");
-		addRecord("Bloggs", "Another description example");
-		addRecord(RandomTestUtil.randomString(), RandomTestUtil.randomString());
-
-		assertSearch("\"Joe Bloggs\"", 1);
-		assertSearch("Bloggs", 2);
-	}
-
-	@Test
-	public void testPunctuationInExactPhrase() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord("Joe? Bloggs!");
-		addRecord("Joe! Bloggs?");
-		addRecord("Joe Bloggs");
-		addRecord("Bloggs");
-
-		assertSearch("\"Joe? Bloggs!\"", 3);
-		assertSearch("\"Joe! Bloggs?\"", 3);
-	}
-
-	@Test
-	public void testQuestionMarksVersusStopwords1() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord(RandomTestUtil.randomString());
-		addRecord("how ? create ? coupon");
-
-		assertSearch("\"how ? create ? coupon\"", 1);
-		assertSearch("\"how to create a coupon\"", 0);
-		assertSearch("\"how with create the coupon\"", 0);
-	}
-
-	@Test
-	public void testQuestionMarksVersusStopwords2() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord(RandomTestUtil.randomString());
-		addRecord("how with create the coupon");
-
-		assertSearch("\"how ? create ? coupon\"", 0);
-		assertSearch("\"how to create a coupon\"", 1);
-		assertSearch("\"how with create the coupon\"", 1);
-	}
-
-	@Test
-	public void testQuestionMarksVersusStopwords3() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord(RandomTestUtil.randomString());
-		addRecord("how to create a coupon");
-
-		assertSearch("\"how ? create ? coupon\"", 0);
-		assertSearch("\"how to create a coupon\"", 1);
-		assertSearch("\"how with create the coupon\"", 1);
-	}
-
-	@Test
-	public void testQuestionMarksVersusStopwords4() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord(RandomTestUtil.randomString());
-		addRecord("how ! create ! coupon");
-
-		assertSearch("\"how ? create ? coupon\"", 1);
-		assertSearch("\"how to create a coupon\"", 0);
-		assertSearch("\"how with create the coupon\"", 0);
-	}
-
-	@Test
-	public void testStopwords() throws Exception {
-		addRecord(RandomTestUtil.randomString());
-		addRecord(RandomTestUtil.randomString(), "Another description example");
-
-		assertSearch("Another The Example", 1);
-	}
-
-	@Test
-	public void testStopwordsInExactPhrase() throws Exception {
-		Assume.assumeTrue(isExactPhraseQueryImplementedForSearchEngine());
-
-		addRecord("how to create a coupon");
-		addRecord("Joe Of Bloggs");
-		addRecord("Joe Bloggs");
-		addRecord("Bloggs");
-
-		assertSearch("\"how to create a coupon\"", 1);
-		assertSearch("\"how with create the coupon\"", 1);
-		assertSearch("\"how Liferay create Liferay coupon\"", 0);
-		assertSearch("\"how create coupon\"", 0);
-		assertSearch("\"Joe Of Bloggs\"", 1);
-		assertSearch("\"Joe The Bloggs\"", 1);
-		assertSearch("\"Bloggs A\"", 3);
-		assertSearch("\"Of Bloggs\"", 3);
-		assertSearch("\"The Bloggs\"", 3);
-	}
-
-	protected static SearchContext getSearchContext(
-			Group group, DDLRecordSet recordSet)
-		throws Exception {
-
-		SearchContext searchContext = SearchContextTestUtil.getSearchContext(
-			group.getGroupId());
-
-		searchContext.setAttribute("recordSetId", recordSet.getRecordSetId());
-		searchContext.setAttribute("status", WorkflowConstants.STATUS_ANY);
-
-		return searchContext;
-	}
-
-	protected void addRecord(String name) throws Exception {
-		addRecord(name, RandomTestUtil.randomString());
-	}
-
-	protected void addRecord(String name, String description) throws Exception {
-		DDMFormValues ddmFormValues = createDDMFormValues();
-
-		DDMFormFieldValue nameDDMFormFieldValue =
-			createLocalizedDDMFormFieldValue("name", name);
-
-		ddmFormValues.addDDMFormFieldValue(nameDDMFormFieldValue);
-
-		DDMFormFieldValue descriptionDDMFormFieldValue =
-			createLocalizedDDMFormFieldValue("description", description);
-
-		ddmFormValues.addDDMFormFieldValue(descriptionDDMFormFieldValue);
-
-		_recordTestHelper.addRecord(
-			ddmFormValues, WorkflowConstants.ACTION_PUBLISH);
-	}
-
-	protected DDLRecordSet addRecordSet() throws Exception {
-		DDLRecordSetTestHelper recordSetTestHelper = new DDLRecordSetTestHelper(
-			_group);
-
-		DDMStructureTestHelper ddmStructureTestHelper =
-			new DDMStructureTestHelper(_group);
-
-		return recordSetTestHelper.addRecordSet(
-			ddmStructureTestHelper.addStructureXsd(this.getClass()));
-	}
-
-	protected void assertSearch(String keywords, int length) {
-		_searchContext.setKeywords(keywords);
-
-		Hits hits = DDLRecordLocalServiceUtil.search(_searchContext);
-
-		Assert.assertEquals(length, hits.getLength());
-	}
-
-	protected DDMFormValues createDDMFormValues() throws Exception {
-		DDLRecordSet recordSet = _recordTestHelper.getRecordSet();
-
-		DDMStructure ddmStructure = recordSet.getDDMStructure();
-
-		return DDMFormValuesTestUtil.createDDMFormValues(
-			ddmStructure.getDDMForm(),
-			DDMFormValuesTestUtil.createAvailableLocales(LocaleUtil.US),
-			LocaleUtil.US);
-	}
-
-	protected DDMFormFieldValue createLocalizedDDMFormFieldValue(
-		String name, String enValue) {
-
-		return DDMFormValuesTestUtil.createLocalizedDDMFormFieldValue(
-			name, enValue);
-	}
-
-	protected boolean isExactPhraseQueryImplementedForSearchEngine() {
-		SearchEngine searchEngine = SearchEngineUtil.getSearchEngine(
-			SearchEngineUtil.getDefaultSearchEngineId());
-
-		String vendor = searchEngine.getVendor();
-
-		if (vendor.equals("Lucene") || vendor.equals("SOLR")) {
-			return true;
-		}
-
-		return false;
-	}
-
-	@DeleteAfterTestRun
-	private Group _group;
-
-	private DDLRecordTestHelper _recordTestHelper;
-	private SearchContext _searchContext;
-
-}
